refactor transfer block to shared and OS-specific
