+++ transfer/block_common.go
@@ -1,7 +1,3 @@
-//go:build linux
-// +build linux
-
-// transfer/block.go
 package transfer
 
 import (
@@ -15,32 +11,11 @@ import (
 	"lvmsync_go/config"
 
 	"go.uber.org/zap"
-	"golang.org/x/sys/unix"
 )
 
 // PipeCreationCount tracks the number of pipes created by ReadBlockWithRetries
 // when no persistent pipe is supplied. It is mainly used for benchmarking.
 var PipeCreationCount int64
-
-func ZeroCopyTransfer(src *os.File, dst *os.File, offset int64, length int64, pipeFds [2]int) error {
-	remaining := length
-	off := offset
-	for remaining > 0 {
-		n, err := unix.Splice(int(src.Fd()), &off, pipeFds[1], nil, int(remaining), 0)
-		if err != nil {
-			return fmt.Errorf("splice read failed: %w", err)
-		}
-		if n == 0 {
-			break
-		}
-		_, err = unix.Splice(pipeFds[0], nil, int(dst.Fd()), nil, int(n), 0)
-		if err != nil {
-			return fmt.Errorf("splice write failed: %w", err)
-		}
-		remaining -= n
-	}
-	return nil
-}
 
 func ReadBlock(src *os.File, offset int64, size int) ([]byte, error) {
 	buf := make([]byte, size)
